Add type annotations for function parameters in multiple files, now it feels more like julia?

+++ fizz_buzz.jl
@@ -1,4 +1,4 @@
-function fizz_buzz(n)
+function fizz_buzz(n::Int64)
     result = []
     for i in 1:n
         if i % 3 == 0 && i % 5 == 0
